+++ backend/sokdak/src/main/java/com/wooteco/sokdak/post/domain/PostDeletionEvent.java
@@ -1,4 +1,4 @@
-package com.wooteco.sokdak.post.event;
+package com.wooteco.sokdak.post.domain;
 
 public class PostDeletionEvent {
