+++ backend/sokdak/src/main/java/com/wooteco/sokdak/post/service/PostService.java
@@ -17,6 +17,7 @@ import com.wooteco.sokdak.member.exception.MemberNotFoundException;
 import com.wooteco.sokdak.member.repository.MemberRepository;
 import com.wooteco.sokdak.member.util.RandomNicknameGenerator;
 import com.wooteco.sokdak.post.domain.Post;
+import com.wooteco.sokdak.post.domain.PostDeletionEvent;
 import com.wooteco.sokdak.post.domain.SearchQuery;
 import com.wooteco.sokdak.post.domain.ViewCountManager;
 import com.wooteco.sokdak.post.dto.NewPostRequest;
@@ -25,7 +26,6 @@ import com.wooteco.sokdak.post.dto.PostDetailResponse;
 import com.wooteco.sokdak.post.dto.PostUpdateRequest;
 import com.wooteco.sokdak.post.dto.PostsCountResponse;
 import com.wooteco.sokdak.post.dto.PostsResponse;
-import com.wooteco.sokdak.post.event.PostDeletionEvent;
 import com.wooteco.sokdak.post.exception.PostNotFoundException;
 import com.wooteco.sokdak.post.repository.PostRepository;
 import java.util.Collections;
