+++ backend/sokdak/src/main/java/com/wooteco/sokdak/report/service/PostReportService.java
@@ -2,7 +2,6 @@ package com.wooteco.sokdak.report.service;
 
 import com.wooteco.sokdak.auth.dto.AuthInfo;
 import com.wooteco.sokdak.auth.service.AuthService;
-import com.wooteco.sokdak.event.NotificationEvent;
 import com.wooteco.sokdak.member.domain.Member;
 import com.wooteco.sokdak.member.exception.MemberNotFoundException;
 import com.wooteco.sokdak.member.repository.MemberRepository;
@@ -67,7 +66,7 @@ public class PostReportService {
 
     private void notifyReportIfOverThanBlockCondition(Post post) {
         if (post.isBlocked()) {
-            applicationEventPublisher.publishEvent(NotificationEvent.toPostReportEvent(
+            applicationEventPublisher.publishEvent(new PostReportEvent(
                     post.getMember().getId(), post.getId()));
         }
     }
